requirements.txt
django>=3.2.1
django-crispy-forms
pillow
requests

opencv-phython
#opencv-python-headless

pylint
pylint-django
